# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Brand {
  id: number;
  brandName: string;
  industry: string;
  branchLocation: string;
  targetAudience: string;
  brandVoice: string;
  /** @nullable */
  websiteUrl: string | null;
  /** @nullable */
  facebookUrl: string | null;
  /** @nullable */
  instagramUrl: string | null;
  /** @nullable */
  tiktokUrl: string | null;
  /** @nullable */
  googlePlaceId: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBrandBody {
  brandName: string;
  industry: string;
  branchLocation: string;
  targetAudience: string;
  brandVoice: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
  /** @nullable */
  googlePlaceId?: string | null;
}

export interface Review {
  id: number;
  brandId: number;
  reviewerName: string;
  rating: number;
  /** @nullable */
  reviewText: string | null;
  reviewDate: string;
  replied: boolean;
  /** @nullable */
  replyText: string | null;
  /** @nullable */
  replyDate: string | null;
  /** @nullable */
  googleReviewId: string | null;
  createdAt: string;
}

export interface CreateReviewBody {
  brandId: number;
  reviewerName: string;
  rating: number;
  /** @nullable */
  reviewText?: string | null;
  reviewDate: string;
  /** @nullable */
  googleReviewId?: string | null;
}

export interface ReviewReplyResponse {
  reply: string;
}

export interface SaveReplyBody {
  replyText: string;
}

export interface ReviewStats {
  brandId: number;
  brandName: string;
  totalReviews: number;
  averageRating: number;
  repliedCount: number;
  unrepliedCount: number;
  rating1Count: number;
  rating2Count: number;
  rating3Count: number;
  rating4Count: number;
  rating5Count: number;
}

export interface GenerateContentBody {
  brandId: number;
  platform: string;
  contentType: string;
  topic: string;
  campaignGoal: string;
  /** @nullable */
  additionalContext?: string | null;
}

export interface GeneratedContent {
  hooks: string[];
  mainCaption: string;
  shortCaption: string;
  cta: string;
  hashtags: string[];
  imagePrompt: string;
  videoPrompt: string;
  visualStyle: string;
}

export interface GenerateStrategyBody {
  brandId: number;
  campaignGoal: string;
  platform: string;
  duration?: string;
}

export interface MarketingStrategy {
  marketingModel: string;
  reasoning: string;
  campaignAngle: string;
  funnelStage: string;
  targetEmotion: string;
  ctaStrategy: string;
  suggestedTopics: string[];
}

export interface ContentPlan {
  id: number;
  brandId: number;
  publishDate: string;
  platform: string;
  contentType: string;
  /** @nullable */
  hook: string | null;
  /** @nullable */
  caption: string | null;
  /** @nullable */
  shortCaption: string | null;
  /** @nullable */
  cta: string | null;
  /** @nullable */
  hashtags: string | null;
  /** @nullable */
  imagePrompt: string | null;
  /** @nullable */
  videoPrompt: string | null;
  topic: string;
  status: string;
  /** @nullable */
  rejectReason: string | null;
  /** @nullable */
  metricoolJobId: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateContentPlanBody {
  brandId: number;
  publishDate: string;
  platform: string;
  contentType: string;
  topic: string;
  /** @nullable */
  hook?: string | null;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  shortCaption?: string | null;
  /** @nullable */
  cta?: string | null;
  /** @nullable */
  hashtags?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  /** @nullable */
  videoPrompt?: string | null;
}

export interface UpdateContentPlanBody {
  publishDate?: string;
  platform?: string;
  contentType?: string;
  topic?: string;
  /** @nullable */
  hook?: string | null;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  shortCaption?: string | null;
  /** @nullable */
  cta?: string | null;
  /** @nullable */
  hashtags?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  /** @nullable */
  videoPrompt?: string | null;
  status?: string;
}

export interface GenerateContentPlanBody {
  brandId: number;
  days: number;
  platform: string;
  campaignGoal: string;
  startDate: string;
}

export interface RejectBody {
  /** @nullable */
  reason?: string | null;
}

export type ListReviewsParams = {
  brandId?: number;
  rating?: number;
  replied?: boolean;
};

export type GetReviewStatsParams = {
  brandId?: number;
};

export type ListContentPlansParams = {
  brandId?: number;
  status?: string;
};