# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Marketing Strategist Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List all brands/stores
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      summary: Create a new brand/store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "201":
          description: Created brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    get:
      operationId: getBrand
      tags: [brands]
      summary: Get brand by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBrand
      tags: [brands]
      summary: Update brand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "200":
          description: Updated brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBrand
      tags: [brands]
      summary: Delete brand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /content/generate:
    post:
      operationId: generateContent
      tags: [content]
      summary: Generate AI marketing content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentBody"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
...
  /content-plans:
    get:
      operationId: listContentPlans
      tags: [contentPlans]
      summary: List content plans
      parameters:
        - name: brandId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of content plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPlan"
    post:
      operationId: createContentPlan
      tags: [contentPlans]
      summary: Create content plan item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentPlanBody"
      responses:
        "201":
          description: Created content plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlan"
  /content-plans/generate:
    post:
      operationId: generateContentPlan
      tags: [contentPlans]
      summary: Generate a 7 or 30-day AI content plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentPlanBody"
      responses:
        "200":
          description: Generated content plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPlan"
  /content-plans/{id}:
    get:
      operationId: getContentPlan
      tags: [contentPlans]
      summary: Get content plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateContentPlan
      tags: [contentPlans]
      summary: Update content plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentPlanBody"
      responses:
        "200":
          description: Updated content plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContentPlan
      tags: [contentPlans]
      summary: Delete content plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /content-plans/{id}/approve:
    post:
      operationId: approveContentPlan
      tags: [approvals]
      summary: Approve content plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlan"
  /content-plans/{id}/reject:
    post:
      operationId: rejectContentPlan
      tags: [approvals]
      summary: Reject content plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlan"
  /content-plans/{id}/publish:
    post:
      operationId: publishContentPlan
      tags: [approvals]
      summary: Publish to Metricool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlan"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Brand:
      type: object
      properties:
        id:
          type: integer
        brandName:
          type: string
        industry:
          type: string
        branchLocation:
          type: string
        targetAudience:
          type: string
        brandVoice:
          type: string
        websiteUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        tiktokUrl:
          type: ["string", "null"]
        googlePlaceId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - brandName
        - industry
        - branchLocation
        - targetAudience
        - brandVoice
        - websiteUrl
        - facebookUrl
        - instagramUrl
        - tiktokUrl
        - googlePlaceId
        - createdAt
        - updatedAt
    CreateBrandBody:
      type: object
      properties:
        brandName:
          type: string
        industry:
          type: string
        branchLocation:
          type: string
        targetAudience:
          type: string
        brandVoice:
          type: string
        websiteUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        tiktokUrl:
          type: ["string", "null"]
        googlePlaceId:
          type: ["string", "null"]
      required:
        - brandName
        - industry
        - branchLocation
        - targetAudience
        - brandVoice
...
    GenerateContentBody:
      type: object
      properties:
        brandId:
          type: integer
        platform:
          type: string
        contentType:
          type: string
        topic:
          type: string
        campaignGoal:
          type: string
        additionalContext:
          type: ["string", "null"]
      required:
        - brandId
        - platform
        - contentType
        - topic
        - campaignGoal
    GeneratedContent:
      type: object
      properties:
        hooks:
          type: array
          items:
            type: string
        mainCaption:
          type: string
        shortCaption:
          type: string
        cta:
          type: string
        hashtags:
          type: array
          items:
            type: string
        imagePrompt:
          type: string
        videoPrompt:
          type: string
        visualStyle:
          type: string
      required:
        - hooks
        - mainCaption
        - shortCaption
        - cta
        - hashtags
        - imagePrompt
        - videoPrompt
        - visualStyle
...
    ContentPlan:
      type: object
      properties:
        id:
          type: integer
        brandId:
          type: integer
        publishDate:
          type: string
          format: date-time
        platform:
          type: string
        contentType:
          type: string
        hook:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        shortCaption:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        videoPrompt:
          type: ["string", "null"]
        topic:
          type: string
        status:
          type: string
        rejectReason:
          type: ["string", "null"]
        metricoolJobId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - brandId
        - publishDate
        - platform
        - contentType
        - hook
        - caption
        - shortCaption
        - cta
        - hashtags
        - imagePrompt
        - videoPrompt
        - topic
        - status
        - rejectReason
        - metricoolJobId
        - createdAt
        - updatedAt
    CreateContentPlanBody:
      type: object
      properties:
        brandId:
          type: integer
        publishDate:
          type: string
          format: date-time
        platform:
          type: string
        contentType:
          type: string
        topic:
          type: string
        hook:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        shortCaption:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        videoPrompt:
          type: ["string", "null"]
      required:
        - brandId
        - publishDate
        - platform
        - contentType
        - topic
    UpdateContentPlanBody:
      type: object
      properties:
        publishDate:
          type: string
          format: date-time
        platform:
          type: string
        contentType:
          type: string
        topic:
          type: string
        hook:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        shortCaption:
          type: ["string", "null"]
        cta:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        imagePrompt:
          type: ["string", "null"]
        videoPrompt:
          type: ["string", "null"]
        status:
          type: string
    GenerateContentPlanBody:
      type: object
      properties:
        brandId:
          type: integer
        days:
          type: integer
        platform:
          type: string
        campaignGoal:
          type: string
        startDate:
          type: string
          format: date-time
      required:
        - brandId
        - days
        - platform
        - campaignGoal
        - startDate
    RejectBody:
      type: object
      properties:
        reason:
          type: ["string", "null"]