# lib/api-zod/src/generated/types/brand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Brand {
  id: number;
  brandName: string;
  industry: string;
  branchLocation: string;
  targetAudience: string;
  brandVoice: string;
  /** @nullable */
  websiteUrl: string | null;
  /** @nullable */
  facebookUrl: string | null;
  /** @nullable */
  instagramUrl: string | null;
  /** @nullable */
  tiktokUrl: string | null;
  /** @nullable */
  googlePlaceId: string | null;
  createdAt: Date;
  updatedAt: Date;
}