# lib/api-zod/src/generated/types/contentPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentPlan {
  id: number;
  brandId: number;
  publishDate: Date;
  platform: string;
  contentType: string;
  /** @nullable */
  hook: string | null;
  /** @nullable */
  caption: string | null;
  /** @nullable */
  shortCaption: string | null;
  /** @nullable */
  cta: string | null;
  /** @nullable */
  hashtags: string | null;
  /** @nullable */
  imagePrompt: string | null;
  /** @nullable */
  videoPrompt: string | null;
  topic: string;
  status: string;
  /** @nullable */
  rejectReason: string | null;
  /** @nullable */
  metricoolJobId: string | null;
  createdAt: Date;
  updatedAt: Date;
}