# lib/api-zod/src/generated/types/createContentPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContentPlanBody {
  brandId: number;
  publishDate: Date;
  platform: string;
  contentType: string;
  topic: string;
  /** @nullable */
  hook?: string | null;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  shortCaption?: string | null;
  /** @nullable */
  cta?: string | null;
  /** @nullable */
  hashtags?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
  /** @nullable */
  videoPrompt?: string | null;
}