# lib/api-zod/src/generated/types/generateContentPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateContentPlanBody {
  brandId: number;
  days: number;
  platform: string;
  campaignGoal: string;
  startDate: Date;
}