# lib/api-zod/src/generated/types/generatedContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedContent {
  hooks: string[];
  mainCaption: string;
  shortCaption: string;
  cta: string;
  hashtags: string[];
  imagePrompt: string;
  videoPrompt: string;
  visualStyle: string;
}