# lib/api-zod/src/generated/types/rejectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Strategist Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RejectBody {
  /** @nullable */
  reason?: string | null;
}